Split the role map and move to relevant components
