+++ src/modules/tccpp/roles.ts
@@ -1,0 +1,6 @@
+import { define_roles } from "../../role-map.js";
+
+export const tccpp_roles = define_roles({
+    historian: { id: "890067617069551646", name: "Historian" },
+    wiki_core: { id: "1354998426370314411", name: "core-wiki" },
+});
