+++ src/modules/wheatley/components/moderation/modmail.ts
@@ -25,7 +25,7 @@ import {
 import { channel_map } from "../../../../channel-map.js";
 import { role_map } from "../../../../role-map.js";
 import { wheatley_channels } from "../../channels.js";
-import { wheatley_roles } from "../../../../roles.js";
+import { wheatley_roles } from "../../roles.js";
 
 /*
  * Flow:
