+++ src/modules/wheatley/components/moderation/rolepersist.ts
@@ -6,7 +6,7 @@ import * as mongo from "mongodb";
 import { capitalize } from "../../../../utils/strings.js";
 import { M } from "../../../../utils/debugging-and-logging.js";
 import { role_map } from "../../../../role-map.js";
-import { wheatley_roles } from "../../../../roles.js";
+import { wheatley_roles } from "../../roles.js";
 import { ModerationComponent, duration_regex } from "./moderation-common.js";
 import { CommandSetBuilder } from "../../../../command-abstractions/command-set-builder.js";
 import {
