+++ src/modules/wheatley/components/redirect.ts
@@ -11,7 +11,7 @@ import { format_list } from "../../../utils/strings.js";
 import { channel_map } from "../../../channel-map.js";
 import { role_map } from "../../../role-map.js";
 import { wheatley_channels } from "../channels.js";
-import { wheatley_roles } from "../../../roles.js";
+import { wheatley_roles } from "../roles.js";
 
 export default class Redirect extends BotComponent {
     private roles = role_map(this.wheatley, wheatley_roles.moderators);
