+++ src/modules/wheatley/components/role-manager.ts
@@ -12,7 +12,7 @@ import { with_retry } from "../../../utils/discord.js";
 import { channel_map } from "../../../channel-map.js";
 import { role_map } from "../../../role-map.js";
 import { wheatley_channels } from "../channels.js";
-import { wheatley_roles } from "../../../roles.js";
+import { wheatley_roles } from "../roles.js";
 
 export type user_role_entry = {
     user_id: string;
